+++ NEWS.md
@@ -1,5 +1,12 @@
 # teal.reporter 0.1.1.9000
 
+### New features
+* Added the collapsible panel for the Show R Code in the previewer.
+* Added additional input when needed to decided if Show R Code should be part of the report.
+
+### Breaking changes
+* Simplified `reporter_previewer_ui` to contain only `id` argument.
+
 # teal.reporter 0.1.1
 
 ### New features