+++ NEWS.md
@@ -1,5 +1,10 @@
 # teal.reporter 0.1.1.9000
 
+### New features
+
+* Updated the `AddCardModule` comment input to have an active cursor when adding a card.
+* Updated report previewer to support preview of multiline comment.
+
 # teal.reporter 0.1.1
 
 ### New features