+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal.reporter 0.1.1.9000
+# teal.reporter 0.1.1.9001
 
 ### New features